notify: use first e-mail address from committers as envelope email
The e-mail provided by GitHub is just the first e-mail that the user set
up in their profile, which is often not the e-mail address that we want
to use. Thus, read the e-mail address from the committers file instead,
if it is available. Use the first address that is in the file for the
specified user.